Automates release process with GitHub Actions
Sets up a comprehensive release workflow using GitHub Actions.
This includes:

- Verifying that the chart and app versions in `Chart.yaml` match the Git tag.
- Building and pushing Docker images to GitHub Container Registry.
- Releasing Helm charts using the `chart-releaser-action`.

Additionally, this commit updates the default image repository and tag in `values.yaml` and modifies `deploy.yaml` to use the chart's appVersion as the default image tag.

Removes the dedicated helm-release workflow in favor of an integrated release process.